pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "SwiftTK"
version = "0.0.1"
authors = [
  { name="Jonathan Harris", email="jono591737@gmail.com" },
]
description = "A wrapper around Tkinter that lets you develop simple multi-page GUI's fast"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "customtkinter",
    "Pillow"
]

[project.urls]
"Homepage" = "https://github.com/yourusername/ctk_router"